Update accessibility-alt-text-bot.yml
### What

The accessibility-alt-text-bot workflow does not have permissions to write a comment on an issue. This is likely due to how permissions are setup in this repo. In order to keep permissions, at the repo level, unchanged, we need to add permission directly to the workflow. 

I also took this as an opportunity to bump the alt-text-bot version and give you access to reminders in discussions