:sparkles: (avatar) - broaden src property, to support complex image-data, eg. StaticImageData interface by next.js
We're basically extending the src prop to also accept objects, which have a `src` key present. This makes sense, as this is the default way of handling images in meta frameworks like next.js. Instead of creating a pseudo type like `ImageData` we could also include the actual types, which would require the installation of new @type dependencies.