:bug: (autocomplete) - fix a bug, where the autocomplete would not emit onChange if the input matches the autocompleteSuggestion
the input value is set to match the suggestion after every input (that is how the autocomplete works). Thus no actual change happens when the incoming key makes the value match the suggestion, as they are already the equal. To fix this issue we have to fire the change event manually on that particular key down