chore: simplify rollup.config.mjs
Since https://github.com/primer/react/pull/6425, the package only exports ESM modules. As a consequence, the rollup configuration does not need to export an array of configuration objects, and as yet another consequence, there is no need to define a base config that could be shared.

Refactor rollup configuration to export a simple default object directly instead of an array.